Update tests and cmdutil
