+++ e2e/testdata/fn-eval/fn-source/exclude-meta-resources-by-default-out-of-place/.expected/exec.sh
@@ -19,7 +19,7 @@ set -eo pipefail
 TEMP_DIR=$(mktemp -d)
 
 kpt fn source\
-| kpt fn eval - --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging \
+| kpt fn eval --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging \
 | kpt fn sink $TEMP_DIR
 
 # copy back the resources
