+++ e2e/testdata/fn-eval/function-chain/.expected/exec.sh
@@ -16,6 +16,6 @@
 set -eo pipefail
 
 kpt fn source \
-| kpt fn eval - --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging \
-| kpt fn eval - --image gcr.io/kpt-fn/set-labels:v0.1.4 -- foo=bar \
+| kpt fn eval --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging \
+| kpt fn eval --image gcr.io/kpt-fn/set-labels:v0.1.4 -- foo=bar \
 | kpt fn sink
