+++ e2e/testdata/fn-eval/function-env/.expected/exec.sh
@@ -19,4 +19,4 @@ IMAGE_TAG="gcr.io/kpt-fn-demo/printenv:v0.1"
 export EXPORT_ENV="export_env_value"
 
 kpt fn source \
-| kpt fn eval - --image $IMAGE_TAG -e EXPORT_ENV -e FOO=BAR
+| kpt fn eval --image $IMAGE_TAG -e EXPORT_ENV -e FOO=BAR
