+++ e2e/testdata/fn-eval/no-image-pull/.expected/exec.sh
@@ -16,7 +16,7 @@
 set -eo pipefail
 
 kpt fn source \
-| kpt fn eval - --image gcr.io/kpt-fn/search-replace:v0.1\
+| kpt fn eval --image gcr.io/kpt-fn/search-replace:v0.1\
 | kpt fn sink .
 
 docker image inspect gcr.io/kpt-fn/search-replace:v0.1
