+++ e2e/testdata/fn-eval/out-of-place-source-sink/.expected/exec.sh
@@ -18,5 +18,5 @@ set -eo pipefail
 rm -rf out; mkdir out
 
 kpt fn source \
-| kpt fn eval - --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging \
+| kpt fn eval --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging \
 | kpt fn sink out
