+++ e2e/testdata/fn-eval/output-to-stdout/.expected/exec.sh
@@ -16,4 +16,4 @@
 set -eo pipefail
 
 kpt fn source |\
-kpt fn eval - --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging
+kpt fn eval --image gcr.io/kpt-fn/set-namespace:v0.1.3 -- namespace=staging
