+++ e2e/testdata/fn-render/out-of-place-fnchain-unwrap/.expected/exec.sh
@@ -3,5 +3,5 @@
 set -eo pipefail
 
 kpt fn render -o stdout \
-| kpt fn eval - --image gcr.io/kpt-fn/set-annotations:v0.1.3 -- foo=bar \
-| kpt fn eval - --image gcr.io/kpt-fn/set-labels:v0.1.3 -o unwrap -- tier=backend
+| kpt fn eval --image gcr.io/kpt-fn/set-annotations:v0.1.3 -- foo=bar \
+| kpt fn eval --image gcr.io/kpt-fn/set-labels:v0.1.3 -o unwrap -- tier=backend
